Flask app, more in the description
This is the flask app, it runs on pythonanywhere. 
It has 7 parameters as input, that do different things. An example call to the URL is below, as well as the description of what the inputs do:
password: needs to be "abcd", or call is rejected
insertStore (boolean), if it is 0, than it simply ignores all other inputs and returns all the details of the store that is in storename. If it is 1, it either updates the store with the other inputs or creates a new store.
StoreID and storeCo are simply fields of the store. The items are added to an itemlist in the store.
searchStore currently is always 0.